+++ changelog.md
@@ -236,3 +236,9 @@
     * Added get_user_home_site
     * Added delete_home_sites
     * Added new_home_site
+
+20-03-2021
+* External Class
+    * Added get_virtual_machines
+* Inventory Class
+    * Added desktop_pool_push_image